# docs/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Tongue API | Documentation",
		"version": "1.0.0",
		"description": "Tongue RESTfull API with CURD",
		"contact": {
			"name": "Marcos Valenzuela",
			"url": "https://marcosvalenzuela.netlify.app/"
		},
		"x-logo": {
			"url": "/favicon.png",
			"altText": "Tongue logo"
		}
	},
...
	"paths": {
...
		"/interactions": {
			"post": {
				"tags": ["Interactions"],
				"summary": "Create a new interaction",
				"description": "Creates a new interaction associated with an existing user and post.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Interaction"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Interaction created successfully."
					},
					"400": {
						"description": "Error occurred while creating the interaction."
					},
					"404": {
						"description": "User or Post not found."
					}
				}
			},
			"get": {
				"tags": ["Interactions"],
				"summary": "Get all interactions",
				"description": "Retrieves a list of all interactions.",
				"parameters": [
					{
						"name": "minify",
						"in": "query",
						"required": false,
						"description": "When set to 'true', returns a list of interactions without populated details for 'user'.",
						"schema": {
							"type": "string",
							"enum": ["true", "false"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "A list of interactions.",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Interaction"
									}
								}
							}
						}
					},
					"400": {
						"description": "Error occurred while fetching the interactions."
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Interaction": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"enum": ["like", "comment"],
						"description": "The type of interaction, which can be either like or comment."
					},
					"content": {
						"type": "string",
						"description": "The content of the interaction, applicable mainly for comments."
					},
					"insertionDate": {
						"type": "string",
						"format": "date-time",
						"description": "The date and time when the interaction was created."
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "The date and time when the interaction was last updated."
					},
					"user": {
						"type": "string",
						"format": "uuid",
						"description": "The ID of the user who made the interaction."
					},
					"post": {
						"type": "string",
						"format": "uuid",
						"description": "The ID of the post that was interacted with."
					}
				},
				"required": ["type", "user", "post"]
			}
		}
	}
}